+++ bin/commands/delete.dart
@@ -1,9 +1,12 @@
-import 'package:mcserv/commands/command.dart';
-import 'package:mcserv/settings/server_chooser.dart';
+import 'package:args/args.dart';
+import 'package:mcserv/settings/settings.dart';
 import 'package:mcserv/settings/settings_helper.dart';
 import 'package:mcserv/utils/utils.dart';
 
-class DeleteCommand extends Command {
+import 'abstract/server_command.dart';
+import 'abstract/yes_flag_mixin.dart';
+
+class DeleteCommand extends ServerCommand with YesFlag {
   @override
   String get name => 'delete';
 
@@ -11,14 +14,15 @@ class DeleteCommand extends Command {
   String get prompt => localizations.deleteCommand;
 
   @override
-  Future<void> execute() async {
-    var server = await chooseServer();
-    if (server == null) {
-      return;
-    }
+  String get description => 'Deletes an existing Server!';
 
-    if (confirm(localizations.confirmDelete, waitForNewLine: true)) {
-      await fs.directory(server.location).delete(recursive: true);
+  @override
+  ArgParser get argParser => withYesFlag(makeArgParser());
+
+  @override
+  Future<void> runOnServer(Installation server) async {
+    if (globalConfirm(localizations.confirmDelete, waitForNewLine: true)) {
+      await findDirectory(server.location).delete(recursive: true);
       await removeServer(server.location);
     }
   }