+++ bin/mcserv.dart
@@ -1,11 +1,13 @@
+import 'dart:async';
 import 'dart:io';
 
 import 'package:args/args.dart';
+import 'package:args/command_runner.dart';
 import 'package:interact/interact.dart';
 import 'package:intl/intl_standalone.dart';
 import 'package:intl/locale.dart';
 import 'package:logging/logging.dart';
-import 'package:mcserv/commands/command.dart';
+import 'package:mcserv/distributions/download.dart';
 import 'package:mcserv/intl/localizations.dart';
 import 'package:mcserv/utils/fs_util.dart';
 import 'package:mcserv/utils/localizations_util.dart';
@@ -14,14 +16,19 @@ import 'package:path/path.dart' as path;
 
 import 'commands/command.dart';
 
+late StreamSubscription _sigIntListener;
+
 Future<ArgResults> _parseArguments(
     ArgParser parser, List<String> arguments) async {
-  final commandNames = allCommands.map((e) => e.name);
-  commandNames.forEach(parser.addCommand);
+  for (var command in allCommands) {
+    parser.addCommand(command.name, command.argParser);
+  }
+  parser.addCommand('help');
 
   parser.addFlag('verbose',
       abbr: 'v', help: localizations.verboseLoggingHelp, negatable: false);
   parser.addFlag('version');
+  parser.addFlag('ascii', negatable: true, defaultsTo: Platform.isWindows);
   parser.addFlag('help',
       abbr: 'h', help: localizations.helpFlagHelp, negatable: false);
   parser.addOption('log-level',
@@ -45,12 +52,12 @@ Never _help(ArgParser parser, {FormatException? e}) {
   exit(1);
 }
 
-Future<void> _initalFlags(ArgParser parser, ArgResults args) async {
-  if (args['help']) {
-    _help(parser);
-  }
+Future<void> _initialFlags(ArgParser parser, ArgResults args) async {
   if (args['version']) {
     await _version();
+  }
+  if (args['ascii']) {
+    Theme.defaultTheme = Theme.basicTheme;
   }
 }
 
@@ -64,10 +71,11 @@ void _initLogger(ArgResults args) {
       print('${record.level.name}: ${record.time}: ${record.message}'));
 }
 
-Command _pickCommand(ArgResults arguments) {
-  final name = arguments.command?.name;
-  if (name != null) {
-    return allCommands.firstWhere((element) => element.name == name);
+ArgResults _pickCommand(
+    Iterable<String> args, ArgParser parser, ArgResults arguments) {
+  final command = arguments.command;
+  if (command != null) {
+    return arguments;
   }
 
   final select = Select(
@@ -75,32 +83,63 @@ Command _pickCommand(ArgResults arguments) {
           options: allCommands.map((e) => e.prompt).toList())
       .interact();
 
-  return allCommands[select];
+  return parser.parse([...args, allCommands[select].name]);
 }
 
 Future<void> _initI18n() async {
   final systemLocale = await findSystemLocale();
 
-  var name =
+  final name =
       systemLocale.length >= 4 ? systemLocale.substring(0, 5) : systemLocale;
-  localizations = await Localizations.load(Locale.parse(name));
+  var locale = Locale.parse("en_US");
+  try {
+    locale = Locale.parse(name);
+  } on FormatException catch (e) {
+    Logger('LocaleLoader').warning('Could not load locale', e);
+  }
+  localizations = await Localizations.load(locale);
 }
 
 void main(List<String> arguments) async {
+  _catchSigint();
   await _initI18n();
   final parser = ArgParser();
   final args = await _parseArguments(parser, arguments);
   _initLogger(args);
-  await _initalFlags(parser, args);
+  await _initialFlags(parser, args);
 
-  final command = _pickCommand(args);
+  final runner = CommandRunner('mcserv', 'mcservdesc');
+  allCommands.forEach(runner.addCommand);
 
-  await command.execute();
+  final commandArgs = _pickCommand(arguments, parser, args);
+
+  await runner.runCommand(commandArgs);
+
+  _close();
+}
+
+void _catchSigint() {
+  // Reset interact cursor to avoid cursor being invisible after SIGINT
+  var sigints = 0;
+  _sigIntListener = ProcessSignal.sigint.watch().listen((event) {
+    if (sigints++ >= 1) {
+      exit(1);
+    } else {
+      _close();
+    }
+  });
+}
+
+void _close() {
+  reset();
+  _sigIntListener.cancel();
+  Download.client.close();
+  closeDio();
 }
 
 Future<void> _version() async {
   final mcServInstall = getInstallationDirectory();
-  final versionFile = fs.file(path.joinAll([...mcServInstall, 'version.txt']));
+  final versionFile = findFile(path.joinAll([...mcServInstall, 'version.txt']));
   Logger('Main').fine('Reading version from ${versionFile.absolute.path}');
   final versionText = await versionFile.readAsString();