+++ lib/utils/confirm.dart
@@ -1,6 +1,13 @@
 import 'package:interact/interact.dart';
 
-bool confirm(String prompt, {defaultValue = false, waitForNewLine = false}) {
+const yesFlag = 'yes';
+
+bool confirm(String prompt,
+    {defaultValue = false, waitForNewLine = false, bool? predefined}) {
+  if (predefined != null) {
+    return predefined;
+  }
+
   final confirm = Confirm(
       prompt: prompt,
       defaultValue: defaultValue,