+++ lib/utils/dio_util.dart
@@ -1,8 +1,17 @@
 import 'package:dio/dio.dart';
 import 'package:logging/logging.dart';
 
+final _dioCache = <Dio>[];
+
+void closeDio() {
+  for (var dio in _dioCache) {
+    dio.close();
+  }
+}
+
 Dio makeDio(Logger log) {
   final dio = Dio();
+  _dioCache.add(dio);
 
   dio.interceptors.add(InterceptorsWrapper(onRequest: (request, handler) {
     log.fine('${request.method} => ${request.uri.toString()}');